it is working

+++ One.Core/DTO/WeatherDto.cs
@@ -8,8 +8,8 @@ namespace One.Core.DTO
     {
         public DateTime Date { get; set; }
         public int Pressure { get; set; }
-        public double FeelLike { get; set; }
+        public decimal FeelLike { get; set; }
         public int Humidity { get; set; }
-        public double Temp { get; set; }
+        public decimal Temp { get; set; }
     }
 }
