+++ One/WeatherMapClient/WeatherMapClientImpl.cs
@@ -23,8 +23,8 @@ namespace One.WeatherMapClient
     }
 
     public class Main    {
-        public double temp { get; set; } 
-        public double feels_like { get; set; } 
+        public decimal temp { get; set; } 
+        public decimal feels_like { get; set; } 
         public double temp_min { get; set; } 
         public double temp_max { get; set; } 
         public int pressure { get; set; }
